+++ CHANGELOG.md
@@ -29,3 +29,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new configuration 'WebConfigPropertyCollections'.
 - Fixed an issue with duplicate resource identifiers in 'WebConfigProperties'.
 - Changed parameter 'Name' to 'Names' in 'WindowsFeatures' Resource according to coding convention.
+- Made the Office Online Server resources actually work and redesigned them.
+  - Added 'OfficeOnlineServerMachineConfig' configuration.